Fix circular dependencies in ng-live-docs
- exported interfaces and injection tokens which caused circular dependency errors
