+++ projects/ng-live-docs/src/documentation-container/documentation-container.component.ts
@@ -5,7 +5,7 @@
 
 import { Component } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { DocumentationEntry } from '../documentation';
+import { DocumentationEntry } from '../interfaces';
 
 @Component({
     selector: 'vmw-documentation-container',
