+++ projects/ng-live-docs/src/ng-live-docs.module.ts
@@ -8,32 +8,11 @@ import { CompoDocRetrieverService } from './compodoc/compodoc-retriever.service'
 import { CompodocSchema } from './compodoc/compodoc-schema';
 import { DocumentationContainerModule } from './documentation-container/documentation-container.module';
 import { DocumentationRetrieverService } from './documentation-retriever.service';
-import { StackBlitzInfo, StackBlitzWriterService, STACKBLITZ_INFO } from './stack-blitz-writer.service';
+import { COMPODOC_URL, DOCUMENTATION_DATA } from './injection-tokens';
+import { StackBlitzInfo } from './interfaces';
+import { StackBlitzWriterService, STACKBLITZ_INFO } from './stack-blitz-writer.service';
 
 const declarations = [];
-
-/**
- * NOTE: The following two has to be exported otherwise the AoT compiler won't see it.
- */
-
-/**
- * Token that makes the documentation JSONs available to the following factory function.
- */
-export const DOCUMENTATION_DATA = new InjectionToken<CompodocSchema[]>(
-    'NgLiveDocsModule.forRoot() CompoDocRetrieverService doc jsons.'
-);
-
-/**
- * A token that is used to provide the url where the raw compodoc output is located.
- */
-export const COMPODOC_URL = new InjectionToken<string>('COMPODOC_URL');
-
-/**
- * Token that makes Stqckblitz JSON data available to factory functions
- */
-export const STACKBLITZ_DATA = new InjectionToken<StackBlitzInfo>(
-    'NgLiveDocsModule.forRoot() StackBlitz template JSON data'
-);
 
 export function getCompoDocRetrieverService(documentations: CompodocSchema[]): DocumentationRetrieverService {
     return new CompoDocRetrieverService(documentations);
